Move Array2d to Lists
